Properly deprecated methods, removed filterTransforms
